ecommerce/requirements/base.txt
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
axios==0.4.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
binary==1.0.1
blinker==1.9.0
boto3==1.37.28
botocore==1.37.28
bs4==0.0.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
collection==0.1.6
colorama==0.4.6
Colr==0.9.1
constantly==23.10.4
contourpy==1.3.1
cors==1.0.1
crypto==1.4.1
cryptography==44.0.2
cryptography-fernet==0.1.0
cssselect==1.3.0
cycler==0.12.1
daphne==4.1.2
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.9
Django==5.2
django-allauth==65.9.0
django-appconf==1.1.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.4
django-cryptography==1.1
django-decouple==2.1
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-fernet-fields==0.6
django-filter==25.1
django-htmx==1.23.0
django-mpesa==2.0.14
django-paypal==2.1
django-phonenumber-field==8.1.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
docutils==0.21.2
dotenv==0.9.9
ecdsa==0.19.1
environ==1.0
fake-useragent==2.1.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.57.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
importlib==1.0.4
incremental==24.7.2
ipaddress==1.0.23
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
JsonForm==0.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
JsonSir==0.0.2
jsonwebtoken==1.0.0
jwt==1.3.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
kombu==5.5.4
logging==0.4.9.6
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mechanize==0.4.10
MouseInfo==0.1.3
msgpack==1.1.0
mss==10.0.0
mysql==0.0.3
mysqlclient==2.2.7
Naked==0.1.32
ngrok==1.4.0
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parsel==1.10.0
passlib==1.7.4
pefile==2024.8.26
phonenumbers==9.0.7
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
Protego==0.4.0
psutil==7.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.2
pydantic_core==2.33.1
PyDispatcher==2.0.7
pydub==0.25.1
pyfiglet==1.0.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.8.1
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pysnmp==7.1.17
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
Python-EasyConfig==0.1.7
python-jose==3.4.0
pytweening==1.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qrcode==8.1
queuelib==1.8.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
Resource==0.2.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.11.4
scapy==2.6.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.31.0
service-identity==24.2.0
setuptools==78.1.0
shellescape==3.8.1
six==1.17.0
skalchemy==0.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.6
speedtest==0.0.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
stripe==12.2.0
termcolor==3.0.1
threadpoolctl==3.6.0
tldextract==5.1.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.30.0
w3lib==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
whois==1.20240129.2
wsproto==1.2.0
xgboost==3.0.0
zope.event==5.0
zope.interface==7.2
